refactor: Commandクラス作成及びdomainクラス移動
refs #002
