+++ MobileSuitDatabase/src/main/java/jp/co/sunarch/mobilesuitDatabase/port/adapter/database/repository/equipment/EquipmentConverter.java
@@ -3,8 +3,8 @@ package jp.co.sunarch.mobilesuitDatabase.port.adapter.database.repository.equipm
 import org.springframework.stereotype.Component;
 
 import jp.co.sunarch.mobilesuitDatabase.domain.model.arms.ArmsId;
-import jp.co.sunarch.mobilesuitDatabase.domain.model.equipment.Equipment;
 import jp.co.sunarch.mobilesuitDatabase.domain.model.mobilesuit.MobileSuitId;
+import jp.co.sunarch.mobilesuitDatabase.domain.model.mobilesuit.Equipment.Equipment;
 import jp.co.sunarch.mobilesuitDatabase.port.adapter.database.repository.equipment.entity.EquipmentEntity;
 
 @Component
